Adding the initial changes for purger fix
