+++ src/main/java/org/opensearch/performanceanalyzer/metrics/PerformanceAnalyzerMetrics.java
@@ -27,7 +27,6 @@
 package org.opensearch.performanceanalyzer.metrics;
 
 
-import java.io.File;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -35,11 +34,9 @@ import java.util.concurrent.ArrayBlockingQueue;
 import java.util.concurrent.BlockingQueue;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.apache.logging.log4j.message.ParameterizedMessage;
-import org.apache.logging.log4j.util.Supplier;
-import org.opensearch.performanceanalyzer.collectors.StatExceptionCode;
-import org.opensearch.performanceanalyzer.collectors.StatsCollector;
+import org.opensearch.performanceanalyzer.PerformanceAnalyzerApp;
 import org.opensearch.performanceanalyzer.config.PluginSettings;
+import org.opensearch.performanceanalyzer.rca.framework.metrics.WriterMetrics;
 import org.opensearch.performanceanalyzer.reader_writer_shared.Event;
 
 @SuppressWarnings("checkstyle:constantname")
@@ -139,7 +136,8 @@ public class PerformanceAnalyzerMetrics {
 
     private static void emitMetric(BlockingQueue<Event> q, Event entry) {
         if (!q.offer(entry)) {
-            // TODO: Emit a metric here.
+            PerformanceAnalyzerApp.WRITER_METRICS_AGGREGATOR.updateStat(
+                    WriterMetrics.METRICS_WRITE_ERROR, "", 1);
             LOG.debug("Could not enter metric {}", entry);
         }
     }
@@ -178,38 +176,6 @@ public class PerformanceAnalyzerMetrics {
             return metricVal.substring(startIndex + 1, endIndex);
         }
         return null;
-    }
-
-    public static void removeMetrics(String keyPath) {
-        removeMetrics(new File(keyPath));
-    }
-
-    public static void removeMetrics(File keyPathFile) {
-        if (keyPathFile.isDirectory()) {
-            String[] children = keyPathFile.list();
-            if (children != null) {
-                for (String child : children) {
-                    removeMetrics(new File(keyPathFile, child));
-                }
-            }
-        }
-        try {
-            if (!keyPathFile.delete()) {
-                // TODO: Add a metric so that we can alarm on file deletion failures.
-                LOG.debug("Purge Could not delete file {}", keyPathFile);
-            }
-        } catch (Exception ex) {
-            StatsCollector.instance().logException(StatExceptionCode.METRICS_REMOVE_ERROR);
-            LOG.debug(
-                    (Supplier<?>)
-                            () ->
-                                    new ParameterizedMessage(
-                                            "Error in deleting file: {} for keyPath:{} with ExceptionCode: {}",
-                                            ex.toString(),
-                                            keyPathFile.getAbsolutePath(),
-                                            StatExceptionCode.METRICS_REMOVE_ERROR.toString()),
-                    ex);
-        }
     }
 
     public static String getJsonCurrentMilliSeconds() {
