+++ src/test/java/org/opensearch/performanceanalyzer/metrics/PerformanceAnalyzerMetricsTests.java
@@ -71,8 +71,8 @@ public class PerformanceAnalyzerMetricsTests {
                 PerformanceAnalyzerMetrics.getMetric(
                         PluginSettings.instance().getMetricsLocation() + "/dir1/test2"));
 
-        PerformanceAnalyzerMetrics.removeMetrics(
-                PluginSettings.instance().getMetricsLocation() + "/dir1");
+        // PerformanceAnalyzerMetrics.removeMetrics(
+        //        PluginSettings.instance().getMetricsLocation() + "/dir1");
     }
 
     // TODO: Turn it on later
